refactor: fetchPurchase 로그인 유저만 가능하게 수정
